add visible filter to projects  - move filters to filters file
